完成商城功能开发与文档更新：修复认证令牌与权限相关问题
- 修复 JWT 密钥变更导致重启后全员掉登录，恢复稳定默认密钥并支持 JWT_SECRET 环境变量注入
- 修复 auth-api 启动失败（JWT_SECRET 占位符补充默认值）
- 修复 BFF Feign 未透传 Authorization 导致管理端权限判定异常
- 新增角色权限管理、区域级联、支付日志、秒杀订单超时补偿等模块
- 同步最新认证与令牌安全经验到 doc/summary.md 及项目文档

+++ goods-store-common/src/main/java/com/fengluan/common/mq/SeckillOrderMessage.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -20,7 +22,7 @@ public class SeckillOrderMessage {
     //会员Id
     private Long memberId;
     //秒杀价格
-    private Long seckillPrice;
+    private BigDecimal seckillPrice;
     //生成的订单号
     private String orderNo;
 }
